Reorganizes project file structure
