+++ VoxelDestroyer.vcxproj
@@ -185,13 +185,15 @@
     <ClInclude Include="cgalhelper.h" />
     <ClInclude Include="constants.h" />
     <ClInclude Include="globalsolver.h" />
-    <ClInclude Include="custommayaconstructs\voxeldragcontext.h" />
-    <ClInclude Include="custommayaconstructs\voxeldragcontextcommand.h" />
-    <ClInclude Include="custommayaconstructs\voxelrendereroverride.h" />
-    <ClInclude Include="custommayaconstructs\voxeldata.h" />
-    <ClInclude Include="custommayaconstructs\particledata.h" />
-    <ClInclude Include="custommayaconstructs\geometryoverride\voxelshape.h" />
-    <ClInclude Include="custommayaconstructs\geometryoverride\voxelsubsceneoverride.h" />
+    <ClInclude Include="custommayaconstructs\tools\voxeldragcontext.h" />
+    <ClInclude Include="custommayaconstructs\tools\voxeldragcontextcommand.h" />
+    <ClInclude Include="custommayaconstructs\data\voxeldata.h" />
+    <ClInclude Include="custommayaconstructs\data\particledata.h" />
+    <ClInclude Include="custommayaconstructs\data\functionaldata.h" />
+    <ClInclude Include="custommayaconstructs\data\d3d11data.h" />
+    <ClInclude Include="custommayaconstructs\draw\voxelrendereroverride.h" />
+    <ClInclude Include="custommayaconstructs\draw\voxelshape.h" />
+    <ClInclude Include="custommayaconstructs\draw\voxelsubsceneoverride.h" />
     <ClInclude Include="directx\directx.h" />
     <ClInclude Include="directx\compute\computeshader.h" />
     <ClInclude Include="directx\compute\vgscompute.h" />
