+++ custommayaconstructs/draw/voxelsubsceneoverride.h
