+++ globalsolver.h
@@ -2,8 +2,8 @@
 
 #include <maya/MPxNode.h>
 #include <maya/MCallbackIdArray.h>
-#include "custommayaconstructs/particledata.h"
-#include "custommayaconstructs/functionaldata.h"
+#include "custommayaconstructs/data/particledata.h"
+#include "custommayaconstructs/data/functionaldata.h"
 #include "directx/compute/dragparticlescompute.h"
 #include "directx/compute/buildcollisiongridcompute.h"
 #include "directx/compute/prefixscancompute.h"
