+++ plugin.cpp
@@ -3,15 +3,15 @@
 #include <maya/MCommandResult.h>
 #include <maya/MFnMessageAttribute.h>
 #include <windows.h>
-#include "custommayaconstructs/voxeldragcontextcommand.h"
+#include "custommayaconstructs/tools/voxeldragcontextcommand.h"
 #include <maya/MAnimControl.h>
 #include <maya/MProgressWindow.h>
-#include "custommayaconstructs/voxeldata.h"
-#include "custommayaconstructs/particledata.h"
-#include "custommayaconstructs/functionaldata.h"
-#include "custommayaconstructs/d3d11data.h"
-#include "custommayaconstructs/geometryoverride/voxelshape.h"
-#include "custommayaconstructs/geometryoverride/voxelsubsceneoverride.h"
+#include "custommayaconstructs/data/voxeldata.h"
+#include "custommayaconstructs/data/particledata.h"
+#include "custommayaconstructs/data/functionaldata.h"
+#include "custommayaconstructs/data/d3d11data.h"
+#include "custommayaconstructs/draw/voxelshape.h"
+#include "custommayaconstructs/draw/voxelsubsceneoverride.h"
 #include <maya/MFnPluginData.h>
 #include <maya/MDrawRegistry.h>
 #include "directx/compute/computeshader.h"
