+++ plugin.h
@@ -23,7 +23,7 @@
 #include <maya/MConditionMessage.h>
 #include <unordered_map>
 #include <string>
-#include "custommayaconstructs/voxelrendereroverride.h"
+#include "custommayaconstructs/draw/voxelrendereroverride.h"
 #include <maya/MViewport2Renderer.h>
 using namespace MHWRender;
